update: setup.go to setup/gest error + improve regex
